+++ konst/src/iter.rs
@@ -36,12 +36,14 @@ pub use collect_const::collect_const;
 pub use self::iter_eval_macro::eval;
 
 #[doc(hidden)]
-pub use self::internal_iter_macros::{__assert_item_ty, __get_item_ty, __items_needs_drop};
+pub use self::internal_iter_macros::{
+    __assert_item_ty, __get_item_ty, __infer_option_ty, __items_needs_drop,
+};
 
 pub use iterator_adaptors::*;
 
 #[doc(hidden)]
-pub use self::iter_eval_macro::__eval2_lowering;
+pub use self::iter_eval_macro::{__eval2_lowering, iter_eval_helpers::__StepByVars};
 
 pub use self::step::Step;