+++ konst/src/iter/iter_eval_macro.rs
@@ -1,3 +1,5 @@
+pub(super) mod iter_eval_helpers;
+
 #[doc(hidden)]
 #[macro_export]
 macro_rules! __iter_eval {
@@ -37,6 +39,24 @@ macro_rules! __iter_eval {
         $crate::__::compile_error!{ "expected iterator argument" }
     );
 }
+
+macro_rules! iter_rev_limitations_docs {
+    () => {
+        concat! {
+            "\n\nIterator-reversing methods can't be called after any of these methods: \n",
+            "- `enumerate` \n",
+            "- `map_while` \n",
+            "- `rev` \n",
+            "- `skip_while` \n",
+            "- `skip` \n",
+            "- `step_by` \n",
+            "- `take_while` \n",
+            "- `take` \n",
+            "- `zip` \n",
+        }
+    };
+}
+pub(super) use iter_rev_limitations_docs;
 
 #[cfg(feature = "cmp")]
 macro_rules! iter_cmp_docs {
@@ -96,6 +116,7 @@ The consuming methods listed alphabetically:
 - [`next`](#next)
 - [`nth`](#nth)
 - [`position`](#position)
+- [`reduce`](#reduce)
 - [`rfind`](#rfind)
 - [`rfold`](#rfold)
 
@@ -139,9 +160,11 @@ The iterator adaptor methods, listed alphabetically
 - [`flat_map`](crate::iter::iterator_dsl#flat_map)
 - [`flatten`](crate::iter::iterator_dsl#flatten)
 - [`map`](crate::iter::iterator_dsl#map)
+- [`map_while`](crate::iter::iterator_dsl#take_while)
 - [`rev`](crate::iter::iterator_dsl#rev)
 - [`skip_while`](crate::iter::iterator_dsl#skip_while)
 - [`skip`](crate::iter::iterator_dsl#skip)
+- [`step_by`](crate::iter::iterator_dsl#step_by)
 - [`take_while`](crate::iter::iterator_dsl#take_while)
 - [`take`](crate::iter::iterator_dsl#take)
 - [`zip`](crate::iter::iterator_dsl#zip)
@@ -308,10 +331,9 @@ assert_eq!(find_parsable(&["10", "20"]), Some(10));
 
 Const equivalent of [`DoubleEndedIterator::rfind`]
 
-Limitations iterator-reversing methods can't:
-- be called more than once in the same macro invocation.
-- be called after calling `take`,`take_while`,`skip`, or `skip_while`.
-
+*/
+#[doc = iter_rev_limitations_docs!()]
+/**
 
 ```rust
 use konst::iter;
@@ -324,6 +346,26 @@ assert_eq!(sum_u64(&[]), None);
 assert_eq!(sum_u64(&[2]), Some(&2));
 assert_eq!(sum_u64(&[2, 5, 8]), Some(&8));
 
+```
+
+### `reduce`
+
+Const equivalent of [`Iterator::reduce`]
+
+```rust
+use konst::iter;
+
+const fn sum_u64(slice: &[u64]) -> u64 {
+    konst::option::unwrap_or!(
+        iter::eval!(slice,copied(),reduce(|accum, rhs| accum + rhs)),
+        0,
+    )
+}
+
+assert_eq!(sum_u64(&[]), 0);
+assert_eq!(sum_u64(&[3]), 3);
+assert_eq!(sum_u64(&[3, 5]), 8);
+assert_eq!(sum_u64(&[3, 5, 8]), 16);
 ```
 
 ### `fold`
@@ -349,8 +391,9 @@ assert_eq!(sum_u64(&[3, 5, 8]), 16);
 
 Const equivalent of [`DoubleEndedIterator::rfold`]
 
-Limitation: iterator-reversing methods can't be called more than once in
-the same macro invocation.
+*/
+#[doc = iter_rev_limitations_docs!()]
+/**
 
 ```rust
 use konst::iter;
@@ -842,7 +885,7 @@ declare_eval2_lowering! {
         $crate::iter::__eval2_lowering!{
             {
                 [$($prev_insts)* (
-                    truncating(skip)
+                    forbids_rev(skip)
                     |_| {
                         let _: $crate::__::usize = skipping;
                         if skipping != 0 {
@@ -867,7 +910,7 @@ declare_eval2_lowering! {
         $crate::iter::__eval2_lowering!{
             {
                 [$($prev_insts)* (
-                    truncating(skip_while)
+                    forbids_rev(skip_while)
                     |ref item| {
                         skipping = skipping && $crate::__parse_closure_1!{
                             ($crate::__eval_closure) (item,) (skip_while),
@@ -894,7 +937,7 @@ declare_eval2_lowering! {
         $crate::iter::__eval2_lowering!{
             {
                 [$($prev_insts)* (
-                    truncating(take)
+                    forbids_rev(take)
                     |_| {
                         let _: $crate::__::usize = taking;
                         if taking == 0 {
@@ -920,7 +963,7 @@ declare_eval2_lowering! {
         $crate::iter::__eval2_lowering!{
             {
                 [$($prev_insts)* (
-                    truncating(take_while)
+                    forbids_rev(take_while)
                     |ref item| {
                         taking = taking && $crate::__parse_closure_1!{
                             ($crate::__eval_closure) (item,) (take_while),
@@ -933,6 +976,48 @@ declare_eval2_lowering! {
                 )]
                 $prev_levels
                 [[is_returning] [$($vars)* mut taking = true,] $iters]
+            }
+
+            $($rem)*
+        }
+    };
+
+    (
+        { [$($prev_insts:tt)*] $prev_levels:tt [$is_returning:tt $vars:tt $iters:tt] }
+        [map_while] ($($closure:tt)*)
+        $($rem:tt)*
+    ) => {
+        $crate::iter::__eval2_lowering!{
+            {
+                [ $($prev_insts)* (map_while $($closure)*)]
+                $prev_levels
+                [[is_returning] $vars $iters]
+            }
+
+            $($rem)*
+        }
+    };
+
+    (
+        { [$($prev_insts:tt)*] $prev_levels:tt [$is_returning:tt [$($vars:tt)*] $iters:tt] }
+        [step_by] ($by:expr $(,)?)
+        $($rem:tt)*
+    ) => {
+        $crate::iter::__eval2_lowering!{
+            {
+                [$($prev_insts)* (
+                    forbids_rev(step_by)
+                    |_| {
+                        vars.left_to_skip -= 1;
+                        if vars.left_to_skip == 0 {
+                            vars.left_to_skip = vars.step;
+                        } else {
+                            continue;
+                        }
+                    }
+                )]
+                $prev_levels
+                [$is_returning [$($vars)* mut vars = $crate::iter::__StepByVars::new($by),] $iters]
             }
 
             $($rem)*
@@ -1229,6 +1314,18 @@ declare_eval2_lowering! {
     };
 
     (
+        { [$($prev_insts:tt)*] $prev_levels:tt $tl_vars:tt }
+        [reduce] ($($closure:tt)*)
+    ) => {
+        $crate::__iter2_finish_lowering!{
+            [accum acc = $crate::__::None]
+            $prev_levels
+            $tl_vars
+            [ $($prev_insts)* (reduce (acc) $($closure)*) ]
+        }
+    };
+
+    (
         {
             [$($prev_insts:tt)*]
             $prev_levels:tt
@@ -1276,7 +1373,7 @@ declare_eval2_lowering! {
             let ord = $crate::iter::__eval2_lowering!{
                 $fixed
                 [__finder __finder] (accum = $crate::__::Equal, |l_item| {
-                    if let Some(r_item) = other_iter.next() {
+                    $crate::if_let_Some!{r_item = other_iter.next() => {
                         $crate::__iter2_require_cmp!{
                             "iterator comparison methods require the  \"cmp\" feature";
                             accum = $crate::cmp::const_cmp!(l_item, r_item);
@@ -1288,7 +1385,7 @@ declare_eval2_lowering! {
                     } else {
                         accum = $crate::__::Greater;
                         __iter2_returner!{}
-                    }
+                    }}
                 })
             };
 
@@ -1310,7 +1407,7 @@ declare_eval2_lowering! {
             let is_eq = $crate::iter::__eval2_lowering!{
                 $fixed
                 [__finder __finder] (accum = true, |l_item| {
-                    if let Some(r_item) = other_iter.next() {
+                    $crate::if_let_Some!{r_item = other_iter.next() => {
                         $crate::__iter2_require_cmp!{
                             "iterator comparison methods require the  \"cmp\" feature";
                             if !$crate::cmp::const_eq!(l_item, r_item) {
@@ -1321,7 +1418,7 @@ declare_eval2_lowering! {
                     } else {
                         accum = false;
                         __iter2_returner!{}
-                    }
+                    }}
                 })
             };
 
@@ -1593,7 +1690,7 @@ macro_rules! declare_eval2_lowering {
                 $fixed:tt
             ) => {
                 $crate::__::compile_error!{$crate::__::concat!{
-                    "expected one of these consumer methods at the end:`",
+                    "expected one of these consumer methods at the end:",
                     $( "\n- ", $crate::__::stringify!($cons_method), )*
                 }}
             };
@@ -1611,6 +1708,8 @@ macro_rules! declare_eval2_lowering {
     )
 }
 use declare_eval2_lowering;
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
 
 #[doc(hidden)]
 #[macro_export]
@@ -1634,7 +1733,7 @@ macro_rules! __iter2_finish_lowering {
 
         // instruction in the current iteration level
         [$($instrs_il:tt)*]
-    ) => { match ($($ret_val,)?) {($(mut $ret_var,)?) => {
+    ) => { match ($($ret_val)?) {($(mut $ret_var)?) => {
         $( let mut $is_returning = false; )?
 
         $(
@@ -1667,6 +1766,8 @@ macro_rules! __iter2_finish_lowering {
     };
 }
 
+////////////////////////////////////////////////////////////////////////////////////////////////////
+
 #[doc(hidden)]
 #[macro_export]
 macro_rules! __iter2_interpreter {
@@ -1688,12 +1789,17 @@ macro_rules! __iter2_interpreter {
     ) => {
         match (
             $($var_val,)*
-            $($iter0_val)? $($item $(@$__use_item)?)?,
-            $($iter_val,)*
+            $crate::__::ManuallyDrop::new($($iter0_val)? $($item $(@$__use_item)?)?),
+            $($crate::__::ManuallyDrop::new($iter_val),)*
         ) {
             ($($var_name,)* iter0, $($iter_name,)*) => {
-                let mut iter0 = $crate::iter::into_iter!(iter0);
-                $(let mut $iter_name = $crate::iter::into_iter!($iter_name);)*
+                let mut iter0 = $crate::iter::into_iter!(
+                    $crate::__::ManuallyDrop::into_inner(iter0)
+                );
+                $(
+                    let mut $iter_name =
+                        $crate::iter::into_iter!($crate::__::ManuallyDrop::into_inner($iter_name));
+                )*
 
                 let elem_phantom_ty = $crate::iter::__get_item_ty(&iter0);
                 $crate::while_let_Some! { $item = iter0.$next() =>
@@ -1766,15 +1872,35 @@ macro_rules! __iter2_interpreter {
         }
     };
 
-    // dedicated instruction for take*/skip* so that rev can detect their presence
+    // dedicated instruction so that rev can detect the presence of incompatible ops
     (
         [$item:ident $next:ident $($__rem_fixed:tt)*]
         $fixed:tt
-        (truncating ($method:ident) |$item_param:pat_param| $code:block)
+        (forbids_rev ($method:ident) |$item_param:pat_param| $code:block)
         $($rem:tt)*
     ) => {
         let $item_param = $item;
         $code
+
+        $crate::__iter2_interpreter!{$fixed $fixed $($rem)*}
+    };
+
+    (
+        [$item:ident $next:ident $($__rem_fixed:tt)*]
+        $fixed:tt
+        (map_while $($closure:tt)*)
+        $($rem:tt)*
+    ) => {
+        let $item = $crate::if_let_Some!{item =
+            $crate::__parse_closure_1!{
+                ($crate::__eval_closure) ($item,) (map_while),
+                $($closure)*
+            } => {
+                item
+            } else {
+                __iter2_returner!{}
+            }
+        };
 
         $crate::__iter2_interpreter!{$fixed $fixed $($rem)*}
     };
@@ -1831,7 +1957,24 @@ macro_rules! __iter2_interpreter {
             $($closure)*
         };
     };
+    (
+        [$item:ident $($__rem_fixed:tt)*]
+        $fixed:tt
+        (reduce ($acc:ident) $($closure:tt)*)
+    ) => {
+        if false {
+            $crate::iter::__infer_option_ty(&$acc, &$item);
+        }
 
+        $acc = $crate::__::Some($crate::if_let_Some!{acc = $acc => {
+            $crate::__parse_closure_2!{
+                ($crate::__eval_closure) ((acc, $item,),) (reduce),
+                $($closure)*
+            }
+        } else {
+            $item
+        }})
+    };
     (
         [$item:ident $($__rem_fixed:tt)*]
         $fixed:tt
@@ -1862,7 +2005,7 @@ macro_rules! __iter2_is_sorted_impl {
                     $crate::__utils::__infer_option_of(&$r_item, &prev)
                 }
 
-                if let $crate::__::Some($l_item) = prev {
+                $crate::if_let_Some!{$l_item = prev => {
                     $crate::__iter2_require_cmp!{
                         "is_sorted methods require the  \"cmp\" feature";
                         if !$comparator {
@@ -1870,7 +2013,7 @@ macro_rules! __iter2_is_sorted_impl {
                             __iter2_returner!{}
                         }
                     }
-                }
+                }}
                 prev = $crate::__::Some($r_item);
             })
         }
@@ -1897,7 +2040,7 @@ macro_rules! __iter2_minmax_impl {
                     $crate::__utils::__infer_option_of(&$r_item, &prev)
                 }
 
-                if let $crate::__::Some($l_item) = prev {
+                $crate::if_let_Some!{$l_item = prev => {
                     $crate::__iter2_require_cmp!{
                         "min/max methods require the  \"cmp\" feature";
                         if let $update_if = $minmaximizer {
@@ -1906,7 +2049,7 @@ macro_rules! __iter2_minmax_impl {
                     }
                 } else {
                     prev = $crate::__::Some($r_item);
-                }
+                }}
 
                 prev
             })
@@ -1982,7 +2125,7 @@ macro_rules! __iter2_rev_asserts {
 #[macro_export]
 macro_rules! __iter2_rev_assert_no_truncation {
     (
-        (truncating ($method_name:ident) $($__0:tt)*)
+        (forbids_rev ($method_name:ident) $($__0:tt)*)
         $($rem:tt)*
     ) => {
         $crate::__::compile_error!{$crate::__::concat!{
@@ -1995,6 +2138,10 @@ macro_rules! __iter2_rev_assert_no_truncation {
     };
     ((map $args:tt enumerate $($_01:tt)*) $($rem:tt)*) => {
         $crate::__::compile_error!{"Cannot call `rev()` after calling `enumerate`"}
+        $crate::__iter2_rev_assert_no_truncation!{ $($rem)* }
+    };
+    ((map_while $($_01:tt)*) $($rem:tt)*) => {
+        $crate::__::compile_error!{"Cannot call `rev()` after calling `map_while`"}
         $crate::__iter2_rev_assert_no_truncation!{ $($rem)* }
     };
     ((zip $($_01:tt)*) $($rem:tt)*) => {