+++ konst/src/iter/iterator_dsl.rs
@@ -31,9 +31,11 @@ The methods listed alphabetically:
 - [`flat_map`](#flat_map)
 - [`flatten`](#flatten)
 - [`map`](#map)
+- [`map_while`](#map_while)
 - [`rev`](#rev)
 - [`skip_while`](#skip_while)
 - [`skip`](#skip)
+- [`step_by`](#step_by)
 - [`take_while`](#take_while)
 - [`take`](#take)
 - [`zip`](#zip)
@@ -146,8 +148,9 @@ assert_eq!(ARR, [2, 4, 6]);
 
 ### `rev`
 
-Limitation: iterator-reversing methods can't be called more than once in
-the same macro invocation.
+*/
+#![doc = crate::iter::iter_eval_macro::iter_rev_limitations_docs!()]
+/*!
 
 ```rust
 use konst::iter;
@@ -155,6 +158,16 @@ use konst::iter;
 const ARR: [&u8; 3] = iter::collect_const!(&u8 => &[2, 3, 5],rev());
 
 assert_eq!(ARR, [&5, &3, &2]);
+```
+
+### `step_by`
+
+```rust
+use konst::iter;
+
+const ARR: [u8; 7] = iter::collect_const!(u8 => 0..20,step_by(3));
+
+assert_eq!(ARR, [0, 3, 6, 9, 12, 15, 18]);
 ```
 
 ### `take`
@@ -177,6 +190,18 @@ const ARR: [&u8; 4] = iter::collect_const!(&u8 =>
 );
 
 assert_eq!(ARR, [&3, &5, &8, &13]);
+```
+
+### `map_while`
+
+```rust
+use konst::iter;
+
+const ARR: [i8; 3] = iter::collect_const!(i8 =>
+    [3i8, 5, 8, 13, 21, 34, 55],map_while(|elem| elem.checked_pow(2) )
+);
+
+assert_eq!(ARR, [9, 25, 64]);
 ```
 
 ### `skip`