+++ konst/tests/misc_tests/array_macros2_ui_tests/drop_type_from_fn_nd_err.rs
@@ -1,0 +1,6 @@
+
+const _: [String; 3] = konst::array::from_fn_nd!(|_| String::new());
+
+
+
+fn main() {}