+++ konst/tests/misc_tests/iter_eval_ui_tests/wrong_types_err.rs
@@ -37,12 +37,20 @@ const fn add(l: u32, r: u32) -> u32 {
     l + r
 }
 
+
 const _: () = {
-    konst::iter::eval!(R, fold(0u32, add));
+    let _: () = konst::iter::eval!(R, reduce(add));
+    let _: () = konst::iter::eval!(R, reduce(|x, y| add(x, y)));
 };
 
 const _: () = {
-    konst::iter::eval!(R, rfold(0u32, |accum, elem| add(accum, elem)));
+    let _: () = konst::iter::eval!(R, fold(0u32, add));
+    let _: () = konst::iter::eval!(R, fold(0u32, |x, y| add(x, y)));
+};
+
+const _: () = {
+    let _: () = konst::iter::eval!(R, rfold(0u32, add));
+    let _: () = konst::iter::eval!(R, rfold(0u32, |accum, elem| add(accum, elem)));
 };
 
 
@@ -60,6 +68,10 @@ const _: () = {
 
 const _: () = {
     konst::iter::eval!(R, rfind(|_| 10u8));
+};
+
+const _: () = {
+    konst::iter::eval!(R, step_by(10u8), next());
 };
 
 const _: () = {